refactor(user): move to auth/domain package
