+++ src/main/java/woowacourse/auth/dao/UserDao.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
-import woowacourse.domain.User;
+import woowacourse.auth.domain.User;
 
 @Repository
 public class UserDao {
